src/backend/requirements/base.txt
Django==4.2.1
dj-rest-auth[with_social]==4.0.1
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
django-cors-headers==3.13.0
whitenoise==6.2.0
httpx==0.24.1
huey==2.4.5
psycopg==3.1.9
redis==4.5.5
